plug create person into readmodel api and test
