+++ server/tests/test_apps/readmodel/test_event_handler.py
@@ -4,6 +4,7 @@ from uuid import uuid4, UUID
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
+from the_history_atlas.apps.domain.core import PersonInput
 from the_history_atlas.apps.domain.models import (
     SummaryAdded,
     SummaryAddedPayload,
@@ -581,3 +582,17 @@ def test_meta_added(engine, DBSession):
 @pytest.mark.xfail(reason="test not implemented")
 def test_meta_tagged(engine, DBSession):
     assert False
+
+
+def test_create_person(readmodel_app, cleanup_tag) -> None:
+    person = PersonInput(
+        wikidata_id="Q1339",
+        wikidata_url="https://www.wikidata.org/wiki/Q1339",
+        name="Johann Sebastian Bach",
+    )
+    created_person = readmodel_app.create_person(person=person)
+    cleanup_tag(created_person.id)
+    assert created_person.id
+    assert created_person.name == "Johann Sebastian Bach"
+    assert created_person.wikidata_id == "Q1339"
+    assert created_person.wikidata_url == "https://www.wikidata.org/wiki/Q1339"
